Update AutoRowHeightLayout.cs
Avoid wasted enumeration of all row heights if the selection is on the header column for which the function is expected to return -1 (at the end of the enumeration). When dealing with large sets (several thousands lines) - the response time for tree sorting through column click is unreasonable due to this enumeration.